Add test cases through a fake Xlib
Implement just enough of xlib to fake the few calls we need to do and
have a bunch of flags in that test lib that can toggle the behavior.
This way we can test all calls for the right handling of the respective
return value, messy though it looks like.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>