graphics-provider: Add an argument to set where the program should be searched
Instead of relying at the environment, with this commit we have an
explicit argument that could be used to set where
`pv_graphics_provider_search_in_path_and_bin()` will look into while it
searches the program.

This gives us more control over which directories we look into.

In the hardcoded common bin directories it is intentionally omitted the current
working directory. That's because it could be a security concern if the cwd
is an attacker-controlled directory.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>