wrap-setup: Keep the XCURSOR_PATH values from the host
If the host has a custom `XCURSOR_PATH` set, we should try to propagate
that into the container too.

This will only work if we already bind-mounted the paths to the same
exact location inside the container.

While this solution doesn't cover all use cases, it should still be
considered an improvement from what we had previously. For example
custom cursors from the home directory should now work as expected.
Additionally this aligns the implementation with what we already do for
`XDG_DATA_DIRS`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>