graphics-provider: Ignore Homebrew paths
The installation guide for Homebrew encourages users to prepend Homebrew
directory in the PATH environment variable.

However, older versions of Homebrew used to install `ldconfig`, and
other low-level tools, into the general Homebrew bin directory that was
usually prepended into PATH. This created some potential issues to
Pressure Vessel because we were unable to find the `ldconfig` that
corresponded to the glibc that we imported into the container.
As it was initially reported in
https://github.com/ValveSoftware/steam-runtime/issues/569

With this commit we skip over all paths that are located under
`/.linuxbrew/` to avoid unexpectedly picking up binaries from Homebrew.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>