wrap-discord: Bind Discord RPC sockets
Discord provides a mechanism called "Rich Presence" that allows games to
synchronize their state with Discord, e.g. showing status information in
the user's profile.

To make this work from inside a Pressure vessel container, we need to
bind-mount the Discord IPC sockets.
They are expected to be called `discord-ipc-`, followed by a number that
ranges from zero up to nine.
They are usually located under `XDG_RUNTIME_DIR`, with `TMPDIR`, `TMP`,
`TEMP` and `/tmp` used as fallback alternatives.

Example games that are known to be using the Discord sockets are:
VRChat, Among Us and osu!

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>