Add display session checks to the system info report
Gather additional information regarding the display server.
This makes it easier to spot issues on end users setup and it will allow
developers differentiate X11 to Wayland sessions.

Checking if a session is using Wayland is not trivial, because the
common environment variable `WAYLAND_DISPLAY` is not necessarily a
requirement. If unset, Wayland will also fallback to its default
`wayland-0` socket. So we need to check if the Wayland socket actually
exists, rather than just looking up to the `WAYLAND_DISPLAY` env var.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>