Add "library_arch" field support for Vulkan ICDs and layer manifests
Layer manifests version 1.2.1 introduced a new field called
"library_arch" that is used to optionally specify the architecture of
the library associated with a manifest.

The same field has also been added for Vulkan driver manifests. In this
case the version was not bumped, so "library_arch" has been included in
the latest version format 1.0.1.

The Vulkan-Loader documentation states that the only supported values
are "32" and "64". However in the loader implementation there isn't a
strict check to validate the "library_arch" value.
For this reason we don't enforce the "32" and "64" values either, as an
attempt to match what the Vulkan-Loader currently does.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>