check-vulkan: Don't try to create two queues in the same family
It's possible (indeed quite likely) that we will choose the same
queue family for the graphics queue (any family with the GRAPHICS_BIT)
and the present queue (any family that supports presentation).
Having more than one VkDeviceQueueCreateInfo for the same family is a
Vulkan spec constraint violation, so we need to uniquify them and only
do one per family. The Nvidia proprietary driver has some logic that
relies on them being unique, leading to a null pointer dereference on
some systems if not.

In the original C++ version of check-vulkan, unique_queue_families was
a std::set (hence its name), but it was wrongly converted into a plain
array that doesn't do anything to ensure they're unique.

steamrt/tasks#294

Fixes: 22d7f70d "check-vulkan: Convert the file from C++ to C"
Signed-off-by: Simon McVittie <smcv@collabora.com>