launch-client: Add the ability to ask to run outside pressure-vessel
This uses IPC from inside the per-app/game container to run commands
outside, and is very similar to `flatpak-spawn --host`. The purpose
of this feature is to let diagnostic tools like MangoHUD run a
command outside the container even though the actual game is running
inside. MangoHUD's initial use-case for this is to run mango_intel_stats
to get Intel GPU statistics. It could also be useful as a way to run
parts of the Steam client, or other client-level features like SteamVR,
inside a container, while still allowing them to run (possibly privileged)
commands on the host system.

This is the reverse of how we have used launcher-service/launch-client
in the past, which was to use IPC from outside the per-app/game container
to run debugging commands inside.

steamrt/tasks#293, steam-runtime-tools#107

Signed-off-by: Simon McVittie <smcv@collabora.com>