pv-adverb: Handle duplicate modules in LD_PRELOAD/LD_AUDIT
If two entries in LD_PRELOAD or LD_AUDIT are literally duplicates of
each other (after syntactic path normalization), then we only need to
create a symlink the first time, and it isn't a problem that we get
EEXIST a second time.

If two entries are not duplicates but have the same basename, recover
by not doing our special handling of LD_PRELOAD/LD_AUDIT modules, and
just using the entry as-is. This might provoke "wrong ELF class"
warnings, but that's no worse than how gameoverlayrenderer.so already
behaves when not using the container runtime.

steamrt/tasks#302

Signed-off-by: Simon McVittie <smcv@collabora.com>