tests: Add a unit test for pv-adverb's LD_AUDIT, LD_PRELOAD handling
The fragility of the dependency handling in
tests/pressure-vessel/meson.build with scout's toolchain continues to be
a mystery to me (especially since Meson wraps the dependency libraries
in --begin-group/--end-group which should make the order irrelevant!).
By experiment, keeping the current order and only linking one of these
libraries seems to be the only thing that works.

Signed-off-by: Simon McVittie <smcv@collabora.com>