launcher-service: Move original environment and cwd from SrtPortalListener
This information was stored in SrtPortalListener, but not for any
particularly good reason, because nothing in that module actually
used it. Move it to the module that makes use of that information.

While doing so, rename original_environ to child_environ, since we
are likely to start making edits to the original environment
for !560.

Signed-off-by: Simon McVittie <smcv@collabora.com>