launcher-service: Add an --alongside-steam option
This allows Steam to provide a process-launching service that is outside
the container runtime (but possibly inside a larger sandbox like Flatpak
or Snap), allowing game processes to run tools such as mango_intel_stats
that cannot usefully be run inside the container (steamrt/tasks#293,
steam-runtime-tools#107).

The intended use is that Steam will run this, with the PATH and
LD_LIBRARY_PATH of the Steam Runtime 1 'scout' environment in effect:

    steam-runtime-launcher-service --alongside-steam

Some defaults change when this option is in use, to be more suitable
for this use-case.

Steam can optionally set environment variable
`SRT_LAUNCHER_SERVICE_ALONGSIDE_STEAM` to
`com.steampowered.PressureVessel.LaunchAlongsideSteam`
in the environment used to launch games, as a way to signal that this
mechanism is available.

If desired, Steam can wait for the launcher service to be ready by passing
a pipe to it as its standard output (fd 1) or as a higher fd referenced by
the --info-fd option, and then reading from that pipe until end-of-file
is reached. If it does this before launching any apps or games, then
apps and games can safely assume that the launcher service is ready for
immediate use.

Alternatively, Steam can let the launcher service start asynchronously,
by passing /dev/null as its standard output. If this is done, an app or
game that starts up very quickly after Steam is launched might find that
`steam-runtime-launch-client --alongside-steam -- ...` does not work
until some time has passed.

To the extent that it's possible, launched processes get the same host
execution environment as the top-level process of Steam itself (the
bin_steam.sh script); this is similar to what happens for non-Steam
app shortcuts. They can re-enter the Steam Runtime 1 'scout' environment,
if desired, by running ~/.steam/root/ubuntu12_32/steam-runtime/run.sh.

Signed-off-by: Simon McVittie <smcv@collabora.com>