tests/input-device: Handle known-failing _srt_evdev_capabilities_guess_type() calls
Our heuristic to identify game controllers, which is essentially the
same as SDL's, cannot distinguish between a 3-axis accelerometer or
gyro (which is not a joystick) and 3-axis driving or simulation pedals
(which are). Before we add test data for those, we should make sure
we can skip those test-cases.

Signed-off-by: Simon McVittie <smcv@collabora.com>