input-monitor: Hex-dump raw evdev capabilities by default
Now that we're using a more readable output format for these, we don't
need to limit them to verbose mode. This gives us essentially the same
information as evemu-describe(1), but without requiring Steam users to
install additional tools.

Because the pretty-printed version only shows known joystick/gamepad
buttons and a few of the most common non-joystick keys, it can be
misleading. For example, we wouldn't show the unassigned button code
used by the 8BitDo SNES30 v0100 via Bluetooth as described in
https://github.com/libsdl-org/SDL/pull/7791, or the multimedia control
keyboard keys found on the Google Stadia Controller rev.A on the same
pull request.

Signed-off-by: Simon McVittie <smcv@collabora.com>