input-device, input-monitor: Add the raw HID report descriptor
This is potentially useful information for any heuristic for identifying
different types of input device, because it is less ambiguous than
the evdev interface. For example, the evdev interface cannot reliably
tell the difference between a 3-axis accelerometer and a set of 3
driving-simulation pedals, because both are often reported as a device
with the same three axes (X, Y, Z) and no buttons.

Signed-off-by: Simon McVittie <smcv@collabora.com>