CI: Use Debian 13 prereleases to test with sanitizers
Now that the tests pass on merged-/usr systems, we can use a more modern
Debian suite.

Normally I'd use the latest Debian stable (Debian 12), but that seems
to have a false positive for a printf argument to %s being NULL,
causing the asan/ubsan builds to fail; so use Debian 10 and 11 as our
representatives of older systems, together with Debian 13 prereleases
as our representative of a modern system.

Signed-off-by: Simon McVittie <smcv@collabora.com>