ld-libs: Make it clearer that /lib:/usr/lib is only a fallback
glibc searches /lib:/usr/lib as a hard-coded fallback, but we don't
expect to get there when searching for dependencies in practice: on
OSs that have dedicated 32- and 64-bit directories (such as
Red-Hat-style multilib, Arch-style multilib and Debian-style multiarch),
those directories are searched indirectly as part of ld.so.cache.

Mentioning that this is a fallback makes it a little bit more obvious
that it isn't an error that we were not searching /usr/lib32, /usr/lib64
or /usr/lib/MULTIARCH when looking for a library of the appropriate
word size.

While I'm here, deduplicate the end of the message a bit.

Suggested-by: Emil Velikov
Signed-off-by: Simon McVittie <smcv@collabora.com>