system-info: Check whether the runtime is on an supportable filesystem
The preferred way to install all versions of the Steam Runtime, both
LD_LIBRARY_PATH and container, is on a fully-featured local Unix
filesystem such as ext4, btrfs or xfs. If that is not how it's installed,
we should flag that in system-info reports so that we can treat it as a
possible root cause for user-visible issues: as with many of our issue
flags, it isn't *necessarily* a problem, but it's a *potential* problem.

Ignore these issue flags during unit testing, because the filesystem
we're running on during unit testing isn't under our control.

Signed-off-by: Simon McVittie <smcv@collabora.com>