check-va-api: Don't claim to have provided a reference frame
We're decoding a single frame, so there is no previous frame that we
could possibly refer to; and our single frame is an i-frame (keyframe),
so we don't need a reference anyway.

When Mesa is compiled with assertions enabled, if num_ref_frames is
greater than 0, it asserts that ReferenceFrames[0] up to
ReferenceFrames[num_ref_frames-1] inclusive are valid handles to
reference frames. In our case, that array is zero-filled, and 0 clearly
can't be a valid handle to a previous frame because we don't have any
previous frames at all, causing an assertion failure and crash.

Closes: steam-runtime-tools#117
Thanks: David Rosca
Signed-off-by: Simon McVittie <smcv@collabora.com>