pv-runtime: Combine multiple resolve_in_sysroot calls
It seems unnecessary to have three almost identical calls to
_srt_file_test_in_sysroot() which each imply a _srt_resolve_in_sysroot()
(differing only in their flags), and potentially a fourth later, when
we could just combine them.

This is not 100% equivalent if we combined more than one of IF_EXISTS,
IF_DIR and IF_REGULAR, but we never actually do that, so keep things
simple by allowing at most one of those flags to be combined. If we find
a use-case for combining them in future, we'll have to decide what that
means.

Signed-off-by: Simon McVittie <smcv@collabora.com>