pv-runtime: When we split manifests per architecture, set library_arch
When we create separate JSON manifests for the x86_64 and i386 versions
of a manifest, we can avoid them being loaded by the loader for the
"other" word size by setting the library_arch field appropriately,
even if it was not set in the original manifest.

For example, if the original manifest said /usr/$LIB/libMangoHud.so,
we'll write out two manifests with library_path set to
/usr/lib/x86_64-linux-gnu/libMangoHud.so and
/usr/lib/i386-linux-gnu/libMangoHud.so. There is no point in trying to
load those into the loader of the opposite word size, so set their
library_arch to "64" and "32" respectively.

Helps: https://github.com/ValveSoftware/steam-runtime/issues/618
Signed-off-by: Simon McVittie <smcv@collabora.com>