pv-runtime: Skip JSON manifests whose library_arch doesn't match
If we find two JSON manifests MangoHud.x86_64.json and MangoHud.x86.json,
with library_arch set to 64 and 32 respectively, but both with
library_path set to /usr/$LIB/libMangoHud.so, then we want to emit
two manifests, one for 64-bit and one for 32-bit. Before this commit, we
would successfully load /usr/$LIB/libMangoHud.so for both word sizes
for each input manifest, and would incorrectly output *four* manifests
in two pairs of duplicates.

Signed-off-by: Simon McVittie <smcv@collabora.com>