pv-adverb: Use g_debug in preference to g_message for environment, etc.
As with commit e7fbf19e "pressure-vessel: Use g_debug for messages that
only show when verbose" in pv-wrap, if opt_verbose is true, then we know
we'll display these even if they are only at debug severity. Using
g_debug instead of g_message gives them the correct prefixes in the log,
and avoids a distracting level of highlighting in the systemd Journal.

Signed-off-by: Simon McVittie <smcv@collabora.com>