pv-runtime: Deduplicate Vulkan drivers, etc. by their manifest content
The unofficial Steam app on Flathub uses an extension to provide i386
graphics drivers, alongside the x86_64 graphics drivers provided by an
extension to the base runtime. Unlike typical Linux distribution
packaging where the JSON manifests are in /usr/share, the JSON manifests
for these drivers are installed into architecture-specific locations.

When using Mesa, this is fine, because each driver's JSON manifest refers
to the Mesa driver by its absolute path, and Vulkan-Loader will ignore
the drivers that are for the other word-size.

However, when using the Nvidia proprietary driver, the JSON manifest
refers to the driver by its basename, which is equally good for both
word-sizes. In a typical Linux distribution, both word-sizes would share
a single JSON manifest and there would still be no duplication, but in
Flatpak, we get two JSON manifests in different locations with identical
content. This has been implicated in crashes in some games, such as
The Elder Scrolls Online (although I have not been able to reproduce this
with the games I've tried).

Before pressure-vessel 0.20230919.0, we accidentally avoided this failure
mode by generating broken symlinks for the i386 manifests in /app, which
meant that only the manifest originally intended to be for x86_64 was
loaded; but that was more a case of compensating errors than anything
else, and was probably breaking the ability to load Mesa drivers for
i386 games (mitigated by the fact that games requiring Vulkan are more
likely to target x86_64). 0.20230919.0 fixed that bug, at the cost of
introducing this duplication for Nvidia drivers.

We can avoid this becoming an issue by only providing the first one of
each set of duplicate JSON manifests.

Resolves: https://github.com/ValveSoftware/steam-runtime/issues/624
Signed-off-by: Simon McVittie <smcv@collabora.com>