log: Lower profiling messages to debug level unless SRT_LOG=timing
There are two reasons we can get profiling messages: either because we
explicitly enabled them with SRT_LOG=timing, or because we have asked to
get all debug messages indiscriminately.

If we're writing to the Journal, g_message() translates into
LOG_NOTICE, which shows up with a distracting level of highlighting.
This is appropriate if we specifically asked for timing information,
but inappropriate if we just want general debug information, so reduce
the log level in the latter case.

Signed-off-by: Simon McVittie <smcv@collabora.com>