pv-runtime: Mount /etc, /var from runtime into interpreter root if appropriate
This generalizes commit 5e0f8520 "pv-runtime: Duplicate symlinks in
/etc, /var in the interpreter root" to use shared machinery.
For symlinks, no functional change is intended; for regular files and
directories, this commit gives them the same handling.

This ensures that readdir() on those paths by an x86 process in FEX-Emu
will show all the files: previously, they would not show up in readdir()
(but, confusingly, directly opening them would still have worked).

This function call cannot legitimately fail, because the only reason
for it to fail would be if the destination was outside /var, which it
is not; so treat failure as a programming error, which will cause a
critical warning.

steamrt/tasks#342

Signed-off-by: Simon McVittie <smcv@collabora.com>