pv-runtime: Create machine ID in interpreter root, too
As with previous commits, this ensures that a readdir() on /etc
or /var/lib by an x86 process under FEX-Emu will show it as existing.

This also reduces repetition a little.

This is actually slightly nicer than what happens under FEX-Emu without
using pressure-vessel, which is that the machine ID only exists in the
real root and does not show up in readdir().

Signed-off-by: Simon McVittie <smcv@collabora.com>