pv-runtime: Bind-mount /etc/passwd, etc. into interpreter root too
As with previous commits, this ensures that they show up in a readdir()
of /etc by an x86 process under FEX-Emu. This commit covers files that
we get from the host system.

As with previous commits, this is actually slightly nicer than what
happens under FEX-Emu without using pressure-vessel, which is that
these files only exist in the real root and do not show up in readdir()
(except for host.conf, which arguably shouldn't be in the rootfs either).

Signed-off-by: Simon McVittie <smcv@collabora.com>