pv-runtime: Use pv_runtime_bind_into_container for files from provider
For files in /etc (and in theory /var), no functional change is intended:
this just converts an open-coded equivalent of
pv_runtime_bind_into_container() into a call into common code.

However, for sockets in /run (as of this commit, only
/run/bumblebee.socket, but others might be added in future), we only
want to populate the real root's /run, not the rootfs' /run; so we need
to distinguish between those two cases.

Signed-off-by: Simon McVittie <smcv@collabora.com>