pv-runtime: Create /etc/localtime, /etc/timezone in interpreter root
As with previous commits, this means they'll show up in a readdir()
from /etc by an x86 process under FEX-Emu.

This is actually slightly nicer than what happens under FEX-Emu without
using pressure-vessel, which is that these files only exist in the
real root and do not show up in readdir().

Signed-off-by: Simon McVittie <smcv@collabora.com>