system-info: Have a SrtSysroot object
This currently assumes that sysroot = `/` is synonymous with using
direct I/O for access to the sysroot. There is no difference yet,
but there will be in a future commit.

This also forces us to be more careful about tracking whether the
sysroot was opened successfully: if it wasn't, self->sysroot will
now be null, and we have to cope with that.

A small functional change in this commit is that the given sysroot was
previously opened with O_NOFOLLOW, so the last path component was
required to be a non-symlink. Now it follows symlinks, including
"magic symlinks" like /proc/self/root.

Signed-off-by: Simon McVittie <smcv@collabora.com>