os: Convert SrtOsRelease into a public object SrtOsInfo
Unlike SrtOsRelease and its accessors in SrtSystemInfo, this keeps track
of everything we might want to add to the JSON report: all os-release(5)
fields (not just the important ones), the source path from which we
got the information, and any diagnostic messages we logged while
collecting it.

Because this is a separate object, it'll be more suitable for
representing the various contexts that might have an os-release(5).

When using FEX-Emu and pressure-vessel, we can have up to four
os-release(5) files: the one that is read by native ARM code, the one
in the interpreter root pseudo-overlayfs that is read by x86 code,
the one that would be read by native ARM code outside the pressure-vessel
container, and the one that would be read by x86 code outside the
pressure-vessel container. Representing these as four objects with the
same API will result in much less duplication.

Signed-off-by: Simon McVittie <smcv@collabora.com>