system-info: Save and load a SrtOsInfo for the container host OS
Previously, s-r-system-info would inspect the live system looking for
host OS information, even if it was meant to be parsing a JSON report
and not looking at any live data.

This needs to be fixed before we can add diagnostic messages to the
OS info part of the JSON report, otherwise diagnostic messages from
failing to find information about the host OS would become part of the
loaded and re-serialized report, causing the round-trip test to fail.

Signed-off-by: Simon McVittie <smcv@collabora.com>