launcher-service: Unblock and unignore all signals during startup
Steam could run this tool in an arbitrary environment, with necessary
signals either blocked (sigprocmask() or similar), or ignored
(sigaction() with action SIG_IGN, or similar). Both of these are
inherited across fork() and execve() by default, and current versions
of Steam do block SIGCHLD in practice.

However, blocking or ignoring SIGCHLD breaks handling of subprocesses,
which is particularly critical for s-r-launcher-service: that prevents
it from detecting that its child processes have exited, causing
s-r-launch-client to hang forever, which in turn breaks startup of
SteamVR on systems where /bin/sh is dash (but not bash, which apparently
resets the disposition of SIGCHLD internally).

steamrt/tasks#354,
https://github.com/ValveSoftware/SteamVR-for-Linux/issues/618

Signed-off-by: Simon McVittie <smcv@collabora.com>