Consistently unblock and unignore signals during CLI tool startup
As with s-r-launcher-service, Steam or a game could run these tools in
an arbitrary environment, with necessary signals either blocked
(sigprocmask() or similar), or ignored (sigaction() with action SIG_IGN,
or similar). Both of these are inherited across fork() and execve() by
default.

s-r-system-info and s-r-check-compatibility do not need this change: they
already called _srt_unblock_signals(), which unblocks all signals, and
unignores selected signals that have been a practical problem in the past.

pv-try-setlocale does not need this change: in practice it only gets
run by pv-adverb, so it inherits pv-adverb's fix for this.

helpers/ do not need this change: they get called with
_srt_child_setup_unblock_signals() and inherit its fix for this.

tests/ do not need this change: Meson should run them in a sensible
environment.

steamrt/tasks#354

Signed-off-by: Simon McVittie <smcv@collabora.com>