utils: Log a warning if we reset a custom signal handler
When our executables get run via execve(), any custom signal
handlers get reset to SIG_DFL by the kernel: nothing else would make
sense, because the pointer to a custom signal handler would no longer be
valid in the new executable's address space. As a result, we start up
with all signals set to either SIG_IGN or SIG_DFL.

_srt_unblock_signals() is meant to be called sufficiently early that
it will not reset any custom signal handlers, because there should not
yet be any. Log a warning if that assumption is broken.

Signed-off-by: Simon McVittie <smcv@collabora.com>