pv-adverb, launcher-service: Use _srt_child_setup_unblock_signals
This avoids some duplication. It does mean that we can't check for
failures, but it's not as if we can really do anything useful about
any error from pthread_sigmask() anyway.

The only errors documented for pthread_sigmask() are EFAULT
(a programming error) and EINVAL (either a programming error, or
use of a signal-set size not supported by the kernel). If we were going
to get EINVAL, we would also have had the same error during program
startup, when we called _srt_unblock_signals(), which *does*
report errors.

Signed-off-by: Simon McVittie <smcv@collabora.com>